Hint_VAD/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
addict=2.4.0=pypi_0
aidisdk=0.12.0=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
black=23.3.0=pypi_0
bleach=6.0.0=pypi_0
boto3=1.26.133=pypi_0
botocore=1.29.133=pypi_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.0=pypi_0
carla=0.9.12=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
clip==0.2.0=pypi_0
cycler=0.11.0=pypi_0
dataclass-type-validator=0.0.7=pypi_0
dataclasses-json=0.5.7=pypi_0
debugpy=1.6.7=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
descartes=1.1.0=pypi_0
dictor=0.1.11=pypi_0
dill=0.3.6=pypi_0
easydict=1.10=pypi_0
entrypoints=0.4=pypi_0
ephem=4.1.4=pypi_0
exceptiongroup=1.1.1=pypi_0
fastjsonschema=2.16.3=pypi_0
filelock=3.12.0=pypi_0
fire=0.5.0=pypi_0
flake8=3.9.2=pypi_0
fonttools=4.38.0=pypi_0
google-auth=2.15.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.54.0=pypi_0
huggingface-hub=0.13.4=pypi_0
idna=3.4=pypi_0
imageio=2.27.0=pypi_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
iniconfig=2.0.0=pypi_0
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.6=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.2.0=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.7=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
llvmlite=0.31.0=pypi_0
lmdb=1.4.0=pypi_0
lyft-dataset-sdk=0.0.8=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=pypi_0
marshmallow=3.19.0=pypi_0
marshmallow-enum=1.5.1=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mccabe=0.6.1=pypi_0
mistune=2.0.5=pypi_0
mmcv-full=1.4.0=pypi_0
mmdet=2.14.0=pypi_0
mmdet3d=0.17.1=pypi_0
mmsegmentation=0.14.1=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclassic=0.5.5=pypi_0
nbclient=0.7.3=pypi_0
nbconvert=7.3.1=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pypi_0
networkx=2.2=pypi_0
notebook=6.5.4=pypi_0
notebook-shim=0.2.2=pypi_0
numba=0.48.0=pypi_0
numpy=1.19.5=pypi_0
nuscenes-devkit=1.1.9=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.1=pypi_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.1=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.1.2=pypi_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=3.2.0=pypi_0
plotly=5.14.1=pypi_0
pluggy=1.0.0=pypi_0
plyfile=0.7.4=pypi_0
prettytable=3.7.0=pypi_0
prometheus-client=0.16.0=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
py-trees=0.8.3=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.6=pypi_0
pycodestyle=2.7.0=pypi_0
pycparser=2.21=pypi_0
pycryptodome=3.17=pypi_0
pydot=1.4.2=pypi_0
pyflakes=2.3.1=pypi_0
pygame=2.3.0=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
pypika=0.48.9=pypi_0
pyquaternion=0.9.9=pypi_0
pyrsistent=0.19.3=pypi_0
pytest=7.3.1=pypi_0
pytest-runner=6.0.0=pypi_0
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.3.0=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.2=pypi_0
qtconsole=5.4.2=pypi_0
qtpy=2.3.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.6.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
send2trash=1.8.0=pypi_0
setuptools=61.2.0=pypi_0
shapely=2.0.1=pypi_0
similaritymeasures=0.7.0=pypi_0
simple-watchdog-timer=0.1.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
sqlparse=0.4.4=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.2=pypi_0
tensorboard=2.9.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.2.0=pypi_0
terminado=0.17.1=pypi_0
terminaltables=3.1.10=pypi_0
threaded=4.1.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.11.2=pypi_0
timm=0.6.12=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=1.9.1+cu111=pypi_0
torchaudio=0.9.1=pypi_0
torchvision=0.10.1+cu111=pypi_0
tornado=6.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
typed-ast=1.5.4=pypi_0
typeguard=2.13.3=pypi_0
typing-extensions=4.5.0=pypi_0
typing-inspect=0.8.0=pypi_0
urllib3=1.26.15=pypi_0
wcwidth=0.2.6=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.5.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=4.0.7=pypi_0
xz=5.2.10=h5eee18b_1
yapf=0.33.0=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
